+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## 0.23.0
 
-- **break**: rename `gro publish` from `gro version` and expand its scope
+- **break**: rename `gro publish` from `gro version` and improve its safety
   ([#189](https://github.com/feltcoop/gro/pull/189))
 - make `config.builds` optional and accept non-array values
   ([#189](https://github.com/feltcoop/gro/pull/189))